gpu: nvgpu: gk20: use u32 timeout for fence wait
Change the gk20_fence_ops wait API to use u32 for the timeout parameter.
The lower APIs called by the wait implementations already use u32, and
all callers were passing in u32s. So this avoids unnecessary casting.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2017505
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gk20a/fence_gk20a.h
@@ -92,7 +92,7 @@ void gk20a_init_fence(struct gk20a_fence *f,
 void gk20a_fence_put(struct gk20a_fence *f);
 struct gk20a_fence *gk20a_fence_get(struct gk20a_fence *f);
 int gk20a_fence_wait(struct gk20a *g, struct gk20a_fence *f,
-							unsigned long timeout);
+							u32 timeout);
 bool gk20a_fence_is_expired(struct gk20a_fence *f);
 int gk20a_fence_install_fd(struct gk20a_fence *f, int fd);
